Crud operation add to each users

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@desc Get All Contacts
 //@route GET /api/contacts
-//@access Public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json(contacts);
 });
 
 //@desc Create New Contacts
 //@route POST /api/contacts
-//@access Public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
   console.log("Request Body:", req.body); // Debugging input
   const { name, email, phone } = req.body;
@@ -20,13 +20,18 @@ const createContact = asyncHandler(async (req, res) => {
     throw new Error("All fields are mandatory!");
   }
 
-  const contact = await Contact.create({ name, email, phone });
+  const contact = await Contact.create({ 
+    name, 
+    email, 
+    phone,
+    user_id: req.user.id
+  });
   res.status(201).json(contact);
 });
 
 //@desc Get  Contacts
 //@route GET /api/contacts
-//@access Public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
@@ -38,12 +43,16 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update Contact
 //@route PUT /api/contacts
-//@access Public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found!");
+  }
+  if (contact.user_id.toString()!== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have persmission to update other user contacts");
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
@@ -58,12 +67,16 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete Contact
 //@route DELETE /api/contacts
-//@access Public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found!");
+  }
+  if (contact.user_id.toString()!== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have persmission to update other user contacts");
   }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
